+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0]
+
+### Added
+- Course addition search support for names (hebrew and english).
+- Dropdown menu for course addition search results.
+- Placeholder text for username, password and ID inputs in options page.
+
+### Changed
+- Updated icons folder names for better clarity.
+- "BGU Site" button in popup page changed to "BGU Sites" and opens list of useful links.
+- Popup Page Year Span input width adjusted for better appearance.
+- Code improvements.
+
+### Fixed
+- Popup page error message not displaying correctly.
+- Options page course name input border not updating correctly.
+
 ## [1.4.2]
 
 ### Fixed